splitted CatalogueSingle into multiple components (refactoring)
