+++ src/assets/styles/_assets.scss
@@ -29,7 +29,7 @@
             width: 100%;
             height: 100%;
             position: relative;
-            // opacity: 0.4;
+            opacity: 0.4;
             &:after {
                 content: '';
                 position: absolute;
@@ -443,6 +443,10 @@
                 background: #f2f2f2;
                 padding: 25px;
                 border-radius: 8px;
+
+                &--centered {
+                    text-align: center;
+                }
             }
             h5 {
                 font-weight: 700;
