+++ src/assets/styles/_btns.scss
@@ -60,6 +60,7 @@
         outline: none;
         box-shadow: none;
         border: none;
+        text-decoration: none !important;
         transition: all 0.4s ease;
         span {
             font-weight: 500;
@@ -67,7 +68,7 @@
     }
 
     &--blue {
-        color: #fff;
+        color: #fff !important;
         background: $secondColor;
     }
